requirements.txt
aioredis==1.2.0
amqp==2.5.1
asgiref==3.2.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.9.3
Automat==0.7.0
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
Django==2.2.5
django-appconf==1.0.3
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-channels==0.7.0
django-image-cropping==1.2.0
django-timezone-field==3.0
django-widget-tweaks==1.4.5
easy-thumbnails==2.6
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.23
incremental==17.5.0
kombu==4.6.5
more-itertools==7.2.0
msgpack==0.6.1
oauthlib==3.1.0
Pillow==6.2.0
pkg-resources==0.0.0
pycparser==2.19
PyHamcrest==1.9.0
pymongo==3.9.0
python-crontab==2.3.9
python-dateutil==2.8.0
pytz==2019.2
redis==3.3.8
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sorl-thumbnail==12.5.0
sqlparse==0.3.0
Twisted==19.7.0
txaio==18.8.1
urllib3==1.25.6
vine==1.3.0
zipp==0.6.0
zope.interface==4.6.0
